back/CraftHub/requirements.txt
asgiref==3.7.2
build==1.0.3
click==8.1.7
colorama==0.4.6
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
packaging==23.2
pillow==10.2.0
pip-tools==7.3.0
PyJWT==2.8.0
pymongo==4.6.1
pyproject_hooks==1.0.0
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.2.4
tzdata==2023.4
wheel==0.42.0
